Do not call std::isalnum() et al. on negative values.
When the buffer contains non-ASCII characters, std::isalnum() and
other similar functions will assert on the Windows debug CRT because
their argument is negative. We avoid this by first casting it to
unsigned char before it gets promoted to int.

+++ src/buffer.cpp
@@ -13,22 +13,22 @@
 namespace
 {
 // A VIM-like definition of a word.  Actually, in Vim this can be changed, but this editor
-// assumes a word is alphanumberic or underscore for consistency
+// assumes a word is alphanumeric or underscore for consistency
 inline bool IsWordChar(const char ch)
 {
-    return std::isalnum(ch) || ch == '_';
+    return std::isalnum((unsigned int)ch) || ch == '_';
 }
 inline bool IsWordOrSepChar(const char ch)
 {
-    return std::isalnum(ch) || ch == '_' || ch == ' ' || ch == '\n' || ch == 0;
+    return std::isalnum((unsigned int)ch) || ch == '_' || ch == ' ' || ch == '\n' || ch == 0;
 }
 inline bool IsWORDChar(const char ch)
 {
-    return std::isgraph(ch);
+    return std::isgraph((unsigned int)ch);
 }
 inline bool IsWORDOrSepChar(const char ch)
 {
-    return std::isgraph(ch) || ch == ' ' || ch == '\n' || ch == 0;
+    return std::isgraph((unsigned int)ch) || ch == ' ' || ch == '\n' || ch == 0;
 }
 inline bool IsSpace(const char ch)
 {
